Update 19110.md
I'm sorry for the multiple commits while I tried to fix this link. I forgot about the 'Preview Changes' option.

Because it ended with a close parentheses, I couldn't make the prior link work in Markdown. It makes me feel a little bit gross to use a shortened url, but at least the link will work. If anybody knows a trick to make the link work, I'll be thankful. The actual URL is below. 
https://geo-ide.noaa.gov/wiki/index.php?title=ISO_19110_(Feature_Catalog)